+++ src/main/java/bean/Comment_SmemberDTO.java
@@ -9,6 +9,10 @@ public class Comment_SmemberDTO {
 	private String comment_writerid;
 	private String comment_content;
 	private Timestamp comment_regdate;
+	private int comment_ref;
+	private int comment_step;
+	private int comment_level;
+	
 	public int getComment_num() {
 		return comment_num;
 	}
@@ -39,8 +43,23 @@ public class Comment_SmemberDTO {
 	public void setComment_regdate(Timestamp comment_regdate) {
 		this.comment_regdate = comment_regdate;
 	}
-	
+	public int getComment_ref() {
+		return comment_ref;
+	}
+	public void setComment_ref(int comment_ref) {
+		this.comment_ref = comment_ref;
+	}
+	public int getComment_step() {
+		return comment_step;
+	}
+	public void setComment_step(int comment_step) {
+		this.comment_step = comment_step;
+	}
+	public int getComment_level() {
+		return comment_level;
+	}
+	public void setComment_level(int comment_level) {
+		this.comment_level = comment_level;
+	}
 	
 }
-	
-
