+++ src/main/java/bean/DTO.java
@@ -1,4 +1,4 @@
-package member.bean;
+package bean;
 
 import java.sql.Timestamp;
