+++ src/main/java/bean/SprojectDTO.java
@@ -25,6 +25,10 @@ public class SprojectDTO {
 	private int good;
 	private int readcount;
 	private Timestamp regdate;
+	private String endProject;
+	private String sent;
+	private String projectName;
+	private int pageNum;
 	
 	
 	public int getNum() {
@@ -159,7 +163,28 @@ public class SprojectDTO {
 	public void setRegdate(Timestamp regdate) {
 		this.regdate = regdate;
 	}
-	
-	
-	
+	public String getEndProject() {
+		return endProject;
+	}
+	public void setEndProject(String endProject) {
+		this.endProject = endProject;
+	}
+	public String getSent() {
+		return sent;
+	}
+	public void setSent(String sent) {
+		this.sent = sent;
+	}
+	public String getProjectName() {
+		return projectName;
+	}
+	public void setProjectName(String projectName) {
+		this.projectName = projectName;
+	}
+	public int getPageNum() {
+		return pageNum;
+	}
+	public void setPageNum(int pageNum) {
+		this.pageNum = pageNum;
+	}
 }
